requirements.txt
django >=3.0, <3.1
django-glrm >=1.1, <1.2
django-markupfield >=2.0, <2.1
django-picklefield >=2.0, <2.1
django-debug-toolbar >=2.2, <2.3
django-countries >=6.1, <6.2
django-mptt >=0.11, <0.12

markdown >=3.0, <3.1
bcoding >=1.5, <1.6
